corr

+++ pointers_arrays_strings/7-print_chessboard.c
@@ -11,11 +11,11 @@
 
 void print_chessboard(char (*a)[8])
 {
-	int i,i2;
+	int i, i2;
 
 	for (i = 0; i < sizeof(a); i++)
 	{
-		for (i2 = 0; i2 < sizeof(a); i2++)
+		for (i2 = 0; i2 < (int) sizeof(a); i2++)
 			printf("%c", a[i][i2]);
 		printf("\n");
 	}
